fix: move useWaitTransaction rq logic into hook
