+++ src/Hooks/useWaitTransaction/useWaitTransaction.ts
@@ -1,42 +1,29 @@
-import { PollExecution } from "@vechain/sdk-errors"
-import { ThorClient, WaitForTransactionOptions } from "@vechain/sdk-network"
-import { useCallback } from "react"
-import { ERROR_EVENTS } from "~Constants"
+import { useQuery } from "@tanstack/react-query"
+import { WaitForTransactionOptions } from "@vechain/sdk-network"
+import { useMemo } from "react"
 import { selectSelectedNetwork, useAppSelector } from "~Storage/Redux"
-import { error } from "~Utils"
+import { waitTransaction } from "./transactionWaiter"
 
-/**
- * Default interval (in ms) for `useWaitTransaction.waitTransaction`
- */
-export const WAIT_TRANSACTION_DEFAULT_INTERVAL = 3000
-/**
- * Default timeout (in ms) for `useWaitTransaction.waitTransaction`. It should be 5 blocks
- */
-export const WAIT_TRANSACTION_DEFAULT_TIMEOUT = 50000
+type UseWaitTransactionOptions = {
+    txId?: string
+} & WaitForTransactionOptions
 
-export const useWaitTransaction = ({
-    intervalMs = WAIT_TRANSACTION_DEFAULT_INTERVAL,
-    timeoutMs = WAIT_TRANSACTION_DEFAULT_TIMEOUT,
-}: WaitForTransactionOptions = {}) => {
+export const useWaitTransaction = (opts: UseWaitTransactionOptions) => {
     const selectedNetwork = useAppSelector(selectSelectedNetwork)
 
-    const waitTransaction = useCallback(
-        async (txId: string) => {
-            try {
-                return await ThorClient.at(selectedNetwork.currentUrl).transactions.waitForTransaction(txId, {
-                    intervalMs,
-                    timeoutMs,
-                })
-            } catch (e) {
-                if (e instanceof PollExecution) {
-                    return null
-                }
-                error(ERROR_EVENTS.SEND, e)
-                throw e
-            }
-        },
-        [intervalMs, selectedNetwork.currentUrl, timeoutMs],
-    )
+    const { data, isFetching } = useQuery({
+        queryKey: ["TransactionReceipt", opts.txId],
+        queryFn: () =>
+            waitTransaction(opts.txId!, {
+                network: selectedNetwork,
+                intervalMs: opts.intervalMs,
+                timeoutMs: opts.timeoutMs,
+            }),
+        enabled: Boolean(opts.txId),
+        retry: false,
+    })
 
-    return { waitTransaction }
+    const isSuccess = useMemo(() => data != null && !data.reverted, [data])
+
+    return { receipt: data, isFetching, isSuccess }
 }
